Add 3 counters and make a few changes on planets.
When player touches the bottom of the planet, the counter will update (+1). Also, I add a layer "Planet" to make sure planets don't collide with planets. (I found that different planets have diffrent falling speed, and if they collide in that way, the one with higher speed will stay on the slower one.)